src/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
python-dotenv
google-generativeai
python-multipart

# Dependências do OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-redis